// src/payload-types.ts
export interface Media {
  id: number;
  /**
   * Required for images (accessibility). For videos, use for accessibility / fallback text.
   */
  alt: string;
  /**
   * Poster/thumbnail for videos. Upload an image to use as preview. Required for videos.
   */
  thumbnail?: (number | null) | Media;
  /**
   * Use "Logos" for company logos (e.g. Home Worked At section). Stores in a separate blob folder.
   */
  prefix?: ("" | "logos" | "works" | "videos" | "images") | null;
  updatedAt: string;
  createdAt: string;
  url?: string | null;
  thumbnailURL?: string | null;
  filename?: string | null;
  mimeType?: string | null;
  filesize?: number | null;
  width?: number | null;
  height?: number | null;
  focalX?: number | null;
  focalY?: number | null;
  sizes?: {
    preview?: {
      url?: string | null;
      width?: number | null;
      height?: number | null;
      mimeType?: string | null;
      filesize?: number | null;
      filename?: string | null;
    };
  };
}
// ...
export interface Work {
  id: number;
  /**
   * e.g. buy-home, car-comparison - must match work page slug
   */
  slug: string;
  title: string;
  description?: {
    root: {
      type: string;
      children: {
        type: any;
        version: number;
        [k: string]: unknown;
      }[];
      direction: ("ltr" | "rtl") | null;
      format: "left" | "start" | "center" | "right" | "end" | "justify" | "";
      indent: number;
      version: number;
    };
    [k: string]: unknown;
  } | null;
  image: number | Media;
  link: {
    href: string;
    target?: ("_self" | "_blank") | null;
  };
  /**
   * Tags to show under the image
   */
  tags?:
    | {
        keyword: string;
        id?: string | null;
      }[]
    | null;
  updatedAt: string;
  createdAt: string;
}